+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The init command now also generate the configuration for tendermint, the flags `
 - [5249](https://github.com/vegaprotocol/vega/issues/5249) - Migrate to Tendermint version 0.35.8
 - [5624](https://github.com/vegaprotocol/vega/issues/5624) - Get rid of `updateFrequency` in price monitoring definition
 - [5601](https://github.com/vegaprotocol/vega/issues/5601) - Remove support for launching a proxy in front of console and token dApp
+- [5872](https://github.com/vegaprotocol/vega/issues/5872) - Remove console and token dApp from networks
 
 ### 🗑️ Deprecation
 - [](https://github.com/vegaprotocol/vega/issues/xxxx) -